+++ src/main.rs
@@ -1,8 +1,10 @@
 #![no_std]
 #![no_main]
+#![feature(abi_x86_interrupt)]
 
 use bootloader::bootinfo::BootInfo;
 
+mod interrupts;
 mod serial;
 mod vga_buffer;
 
@@ -17,9 +19,14 @@ bootloader::entry_point!(kernel_main);
 
 fn kernel_main(boot_info: &'static BootInfo) -> ! {
     vga_buffer::init(boot_info.physical_memory_offset);
+    interrupts::init_idt();
 
     println!("[SYSTEM] Neural Microkernel Iniciado. Aguardando integracao NPU/Ring 0.");
     serial_println!("[SYSTEM] Neural Microkernel Iniciado. Aguardando integracao NPU/Ring 0.");
+
+    println!("[TEST] Forcando Breakpoint (int3)...");
+    serial_println!("[TEST] Forcando Breakpoint (int3)...");
+    x86_64::instructions::interrupts::int3();
 
     loop {}
 }